+++ ase/spacegroup/spacegroup.py
@@ -467,7 +467,7 @@ class Spacegroup:
                [ 0.,  0.,  0.]])
         """
         scaled = np.array(scaled_positions, ndmin=2)
-        normalised = np.empty(scaled.shape, np.float)
+        normalised = np.empty(scaled.shape, float)
         rot, trans = self.get_op()
         for i, pos in enumerate(scaled):
             sympos = np.dot(rot, pos) + trans
@@ -633,7 +633,7 @@ def _read_datafile_entry(spg, no, symbol, setting, f):
     spg._scaled_primitive_cell = np.array(
         [[float(floats.get(s, s)) for s in f.readline().split()]
          for i in range(3)],
-        dtype=np.float)
+        dtype=float)
     # primitive reciprocal vectors
     f.readline()
     spg._reciprocal_cell = np.array([[int(i) for i in f.readline().split()]
@@ -644,12 +644,12 @@ def _read_datafile_entry(spg, no, symbol, setting, f):
     spg._subtrans = np.array(
         [[float(floats.get(t, t)) for t in f.readline().split()]
          for i in range(spg._nsubtrans)],
-        dtype=np.float)
+        dtype=float)
     # symmetry operations
     nsym = int(f.readline().split()[0])
     symop = np.array([[float(floats.get(s, s)) for s in f.readline().split()]
                       for i in range(nsym)],
-                     dtype=np.float)
+                     dtype=float)
     spg._nsymop = nsym
     spg._rotations = np.array(symop[:, :9].reshape((nsym, 3, 3)), dtype=int)
     spg._translations = symop[:, 9:]
